requirements.txt
flask~=3.0.3
requests~=2.32.3
pillow~=9.4.0